Preparing Bricks for PR
- Cleaning the build from two separate binaries
- Adding code that runs in Bricks for testing
- C code is only user mode now
- Refactoring both Bricks and Tychools

+++ crates/bricks/src/bricks_const.rs
@@ -1,4 +1,4 @@
-pub const SUCCESS: u32 = 0;
-pub const FAILURE: u32 = 1;
+pub const SUCCESS: u64 = 0;
+pub const FAILURE: u64 = 1;
 
 pub const RET_CODE_BYTES: u64 = 8;
